requirements.txt
# LLM呼び出しとチャットボット用
openai>=0.27.0
streamlit>=1.16.0

# 自然言語処理 (NLP) 用
nltk>=3.6.5

# RAGシステム向け（ベクトルDB構築用）
faiss-cpu>=1.7.2
numpy>=1.21.0

# データ処理と機械学習用
pandas>=1.3.0
scikit-learn>=1.0.0
joblib>=1.1.0

# LangChainなど外部プラットフォーム連携用
langchain>=0.0.142

# 環境変数管理（必要に応じて）
python-dotenv>=0.20.0
